Update the explanation of GHGSat sensor
